% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_rtestimate.R
\name{new_summarize}
\alias{new_summarize}
\title{Create a new summary object}
\usage{
new_summarize(date, median, lb, up, package)
}
\arguments{
\item{date}{Integer vector. vector of index dates.}

\item{median}{Double vector. vector of median values.}

\item{lb}{Double vector. vector of lower bounds.}

\item{up}{Double vector. vector of upper bounds.}

\item{package}{String. Name of the package.}
}
\description{
Create a new summary object
}
