% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_rtestimate.R
\name{summarize_rtestimate}
\alias{summarize_rtestimate}
\alias{summarize_rtestimate.default}
\alias{summarize_rtestimate.cv_poisson_rt}
\title{Extract Rt estimation from a model fit}
\usage{
summarize_rtestimate(x, ...)

\method{summarize_rtestimate}{default}(x, ...)

\method{summarize_rtestimate}{cv_poisson_rt}(x, level = 0.95, lambda, ...)
}
\arguments{
\item{x}{Object to extract Rt from.}

\item{...}{Additional arguments passed to methods.}

\item{level}{Confidence level for the confidence interval.}

\item{lambda}{The Poisson parameter (\code{cv_poisson_rt}).}
}
\description{
Extract Rt estimation from a model fit
}
