% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_rtestimate.R
\name{new_summrt}
\alias{new_summrt}
\alias{print.summrt_summary}
\title{Create a new summary object}
\usage{
new_summrt(date, median, lb, ub, package, notes)

\method{print}{summrt_summary}(x, ...)
}
\arguments{
\item{date}{Integer vector. vector of index dates.}

\item{median}{Double vector. vector of median values.}

\item{lb}{Double vector. vector of lower bounds.}

\item{ub}{Double vector. vector of upper bounds.}

\item{package}{String. Name of the package.}

\item{notes}{String. Notes about the summary.}

\item{x}{An object of class \code{summrt_summary}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A list of class \code{summrt_summary}. with the following components:
\itemize{
\item \code{estimates}: A tibble with the following columns:
\itemize{
\item \code{date}: Integer vector. vector of index dates.
\item \code{median}: Double vector. vector of median values.
\item \code{lb}: Double vector. vector of lower bounds.
\item \code{ub}: Double vector. vector of upper bounds.
}
\item \code{package}: String. Name of the package.
\item \code{notes}: String. Notes about the summary.
}
}
\description{
Creates a new summary object for the \code{summrt} package while validating the input.
}
