% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_rtestimate.R
\name{summarize_rtestimate}
\alias{summarize_rtestimate}
\alias{summarize_rtestimate.default}
\alias{summarize_rtestimate.cv_poisson_rt}
\alias{summarize_rtestimate.poisson_rt}
\alias{summarize_rtestimate.epinow}
\alias{summarize_rtestimate.estimate_R}
\alias{summarize_rtestimate.Rt}
\title{Extract Rt estimation from a model fit}
\usage{
summarize_rtestimate(x, ...)

\method{summarize_rtestimate}{default}(x, ...)

\method{summarize_rtestimate}{cv_poisson_rt}(x, level = 0.95, lambda = "lambda.1se", ...)

\method{summarize_rtestimate}{poisson_rt}(x, level = 0.95, lambda = NULL, ...)

\method{summarize_rtestimate}{epinow}(x, level = 0.95, ...)

\method{summarize_rtestimate}{estimate_R}(x, ...)

\method{summarize_rtestimate}{Rt}(x, ...)
}
\arguments{
\item{x}{Object to extract Rt from.}

\item{...}{Additional arguments passed to methods.}

\item{level}{Confidence level for the confidence interval.}

\item{lambda}{The Poisson parameter (\code{cv_poisson_rt}).}
}
\description{
Extract Rt estimation from a model fit
}
\details{
The \code{estimate_R} method is for the \code{EpiEstim} package.

The \code{Rt} method is for the \code{EpiLPS} package.
}
